json字符串保存或更新

+++ unifystorage_core/src/main/java/com/winning/unifystorage_core/Handler/DeleteHandler.java
@@ -72,12 +72,15 @@ public class DeleteHandler extends HandlerAdapter {
             result.addChangeListener(new OrderedRealmCollectionChangeListener<RealmResults>() {
                 @Override
                 public void onChange(RealmResults realmResults, OrderedCollectionChangeSet changeSet) {
-                    if (realmResults.size() > 0) {
-                        deleteByQueryTransaction(realmResults);
-                    }else {
-                        dbResult.setCount(0);
-                        dbResult.setResultCallback(false,new Throwable("No data was selected"));
+                    if (OrderedCollectionChangeSet.State.INITIAL == changeSet.getState()){
+                        if (realmResults.size() > 0) {
+                            deleteByQueryTransaction(realmResults);
+                        }else {
+                            dbResult.setCount(0);
+                            dbResult.setResultCallback(false,new Throwable("No data was selected"));
+                        }
                     }
+
                 }
             });
